wsd: test: merge UnitWopiUnlock into UnitWOPILock.cpp
